update : rename domain to entity
