+++ src/main/java/f1b4/webide_server/entity/code/JavaCodeUtil.java
@@ -1,4 +1,4 @@
-package f1b4.webide_server.domain.code;
+package f1b4.webide_server.entity.code;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
